+++ Assets/Scripts/Cgs/CardGameView/Multiplayer/CgsNetPlayable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd0ce7e045d8ad3419bdf9c4ceab86be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
